+++ docker/srcs/uwsgi-django/pong/views/navi_views.py
@@ -27,11 +27,6 @@ def pong_view(request):
 		return render(request, 'pong/index.html', param)
 	return render(request, 'pong/index.html')
 
-# def results(request):
-# 	results_list = PongGameResult.objects.all().order_by("-date")
-# 	context = {'results_list': results_list}
-# 	return render(request, 'pong/results.html', context)
-
 # 3D
 def play_tournament(request, match_id):
 	# match_idが整数でない場合の遷移
@@ -71,35 +66,3 @@ def play_tournament(request, match_id):
 def game(request):
 	return render(request, 'pong/game.html')
 
-
-# ーーーーーーーーーーーーーーー
-# old
-# ーーーーーーーーーーーーーーー
-
-# @csrf_exempt #開発用のおまじない
-# def save_game_result(request):
-# 	if request.method == 'POST':
-# 		try:
-# 			data = json.loads(request.body.decode('utf-8'))  # 受け取ったJSONデータをPythonの辞書に変換
-# 			match_id = data.get('match_id')
-# 			player_1_score = data.get('player_1_score')
-# 			player_2_score = data.get('player_2_score')
-
-# 			# 勝者を決定
-# 			winner = data.get('player_1_name') if player_1_score > player_2_score else data.get('player_2_name')
-# 			loser = data.get('player_2_name') if player_1_score > player_2_score else data.get('player_1_name')
-
-# 			# PongGameResult インスタンスを作成して保存
-# 			game_result = PongGameResult(
-# 				match_id=match_id,
-# 				player_1_score=player_1_score,
-# 				player_2_score=player_2_score,
-# 				name_winner=winner,
-# 				# name_loser=loser,
-# 			)
-# 			game_result.save()
-
-# 			return JsonResponse({'status': 'success'})
-# 		except Exception as e:
-# 			return JsonResponse({'status': 'error', 'message': str(e)})
-# 	return JsonResponse({'status': 'error', 'message': 'Invalid request'}, status=400)
